Adds npm install to readme
Problem
The README does not currently show how to install this sdk in a project.
Adding this line to the README allows developers to easily install this
sdk without having to find it in the npm registry. Developers can now
copy this easily from the README on github!

Solution

I've added a code block with the proper npm registry install link! This
was found at https://www.npmjs.com/package/@spotify/web-api-ts-sdk